Cleanup sync_rules id sequence usage. Use integer/number type for group_id throughout.

+++ modules/module-postgres-storage/src/migrations/scripts/1684951997326-init.ts
@@ -47,7 +47,7 @@ export const up: migrations.PowerSyncMigrationFunction = async (context) => {
 
     await db.sql`
       CREATE TABLE sync_rules (
-        id BIGSERIAL PRIMARY KEY,
+        id INTEGER PRIMARY KEY,
         state TEXT NOT NULL,
         snapshot_done BOOLEAN NOT NULL DEFAULT FALSE,
         last_checkpoint BIGINT,
